Extract artifact resolver & file resolver methods into utility classes.
